수업 & 보유과목추가 그 외 등등..

+++ src/main/java/com/ezen/management/dto/SubjectHoldDTO.java
@@ -17,6 +17,6 @@ public class SubjectHoldDTO {
 
     private String name;
 
-    private int lesson_idx;
+    private Long lesson_idx;
 
 }
